refactor : 예약페이지-라우팅-수정
